+++ app/src/main/java/com/example/playlist_maker/music_library/domain/FavoriteTracksInteractor.kt
@@ -10,5 +10,5 @@ interface FavoriteTracksInteractor {
 
     suspend fun deleteTrack(trackId: Int)
 
-    suspend fun isFavoriteTrack(trackId: Long): Flow<Boolean>
+    suspend fun isFavoriteTrack(trackId: Int): Flow<Boolean>
 }
